Changed truefans backend to separate database.
Completely changed the backend database functionality. Now truefans
is a separate database (class.TrueFansDb.php that is opened by a pdo
object with it's own DB credentials. Previously truefans was
simply a table in MW but I separated them to allow for integration
with Drupal in the future. class.TrueFansDb.php may also be integrated
into some other backend like CivicCRM and this would also require
a separate database. SpecialPage was also updated to reflect this.