Failed attempt to recreate paypal form as object
Commiting this for future reference. I'm unable to make the paypal
form a normal object because HTMLFormField prefixes 'wp' to all the
names which are required by remote paypal server. There are no
methods to change this so the paypal form now must be included
as raw html. C'est la vie.