Cleaned code and resolved request system ungliness.
I got the request system to work sanely. We can load hidden variables
autmatically with forms -- and forms load their requests as desired.
A post callback function was implemented in the TrueFanForm class
which also eliminated the call_usr_func() bug that taunted me for a
day. The javascript that adds EmailInput fields also works mostly
as desired. Fake hidden fields were eliminated which cleaned out
some css. Now hidden form fields *that can autoload* are implemented
with custom HTMLPassableHidden class which simply adds the correct
name to field when it is output as html.