Eliminated hidden variables.
Despite significant effort to get the hidden variables to work,
I got rid of them since they provide redundant information once
the idea of verifying against CSRF is taken into consideration.
Now the a true fan profile is loaded as a member variable of the
page and reloaded if a post causes editing. Thus DB requests
for profiles are limited to 0,1, or 2 request.