Added the ability to email.
An extra button to submit emails was added to the edit
form. This will certainly be changed later, but the
functionality is their...Added the ability to change name
upon initial submission. Haven't yet added that to the edit
form. Considering just a single update function for the
edit form that updates all sent fields and doens't make
extra submits to the database.