Code quality spit shine
Significant effort was put forth to increase the quality of the
code and make more permanent choices about the code base. Tons
of rearragnement but we shall all be more sane for it. Cleaned
out lots of globals, although some reside remains. Decent amount
of usability improvement, but more of that on the next commit.